Disable zoom as a temporary fix for Chrome disambiguation pop ups
Zoom should be enabled, and I'm recording this disabling of zoom as a bug. Currently this is the only way I can find to prevent Chrome on Android (mobile in general?) from displaying a zoomed popup every time a board tile is touched, as Chrome considers the tiles too small to be unambiguously touched and I cannot override this decision without removing the option to zoom in general.